22082敵ユニットをDestroyした時のエラー解消、レイアウト中に攻撃用コライダーにオブジェクトが衝突した時に出るエラーの解消、ノックバック演出の追加

+++ Assets/Script/UnitData.cs
@@ -7,7 +7,7 @@ public class UnitData
     public int cost;
     public int hp;
     public int attackPower;
-    public int blowPower;
+    public float blowPower;
     public float moveSpeed;
     public float weight;
     public float intervalTime;
